[refactor]나의 프로젝트 패키지명 변경 #68
