+++ src/main/java/wowmarket/wow_server/mypage/myproject/MySalesOrder/dto/MySalesOrderDetailDto.java
@@ -1,4 +1,4 @@
-package wowmarket.wow_server.mypage.myproject.mysalesorder.dto;
+package wowmarket.wow_server.mypage.myproject.MySalesOrder.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
