+++ src/main/java/wowmarket/wow_server/mypage/myproject/MySalesOrder/dto/MySalesOrderListResponseDto.java
@@ -1,8 +1,7 @@
-package wowmarket.wow_server.mypage.myproject.mysalesorder.dto;
+package wowmarket.wow_server.mypage.myproject.MySalesOrder.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import wowmarket.wow_server.mypage.myproject.mysalesorder.dto.MySalesOrderDto;
 
 import java.util.List;
