+++ src/main/java/wowmarket/wow_server/mypage/myproject/MySalesProject/dto/MySalesListResponseDto.java
@@ -1,4 +1,4 @@
-package wowmarket.wow_server.mypage.myproject.mysales.dto;
+package wowmarket.wow_server.mypage.myproject.MySalesProject.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
